apps/azure_functions/ri_collector/requirements.txt
azure-functions==1.14.0
azure-identity==1.14.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.16.0
azure-eventhub==5.11.1
beautifulsoup4==4.12.2
feedparser==6.0.10
openai==1.10.0
requests==2.31.0
psycopg2-binary==2.9.6
numpy==1.23.5
pandas==1.5.3
fastavro==1.7.4
brapi
python-dotenv==1.0.0
pytest==8.2.1
pytest-cov==4.1.0
unittest2==1.1.0
mock==5.1.0
black==24.4.0
cryptography==3.4.8
httpx==0.24.0


